# uploads/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SpinalCore Graph API",
    "version": "1.0.0",
    "description": "Welcome to the reference documentation for the Spinalcore Digital Twin REST API! </br></br>To learn about the common use cases and concept of Spinalcore REST APIs, check out our resource center  (https://resourcecenter.fr.spinalcom.com/ressources-d%C3%A9veloppeur).</br></br>In addition to Spinalcore API Reference, we also provide complete documentation for integrator that need to install and implement Spinalcore Digital Twin Operating System on their built environment (https://resourcecenter.fr.spinalcom.com/ressources-int%C3%A9grateur)",
    "termsOfService": "",
    "contact": {
      "email": "developers@spinalcom.com"
    },
    "x-logo": {
      "url": "/logo.png"
    },
    "x-preferred": true,
    "x-providerName": "spinalcom.com",
    "x-unofficialSpec": true
  },
  "components": {
    "securitySchemes": {
      "OauthSecurity": {
        "type": "oauth2",
        "description": "This API uses OAuth 2 with the implicit grant flow. [More info](https://api.example.com/docs/auth)",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "http://localhost:8080/oauth2/token",
            "scopes": {
              "read": "read request",
              "admin": "admin request"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/equipementsGroup/{contextId}/category/{categoryId}/read": {
      "get": {
        "security": [
          {
            "OauthSecurity": ["read"]
          }
        ],
        "description": "read category equipements Group",
        "summary": "Get category equipements Group",
        "tags": ["Equipements Group"],
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "path",
            "name": "categoryId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/BasicNode"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/v1/equipementsGroup/{contextId}/category/{categoryId}/update": {
      "put": {
        "security": [
          {
            "OauthSecurity": ["read"]
          }
        ],
        "description": "update category equipements Group",
        "summary": "update category equipements Group",
        "tags": ["Equipements Group"],
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "path",
            "name": "categoryId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["newNameCategory", "newNameIcon"],
                "properties": {
                  "newNameCategory": {
                    "type": "string"
                  },
                  "newNameIcon": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update Successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/api/v1/equipementsGroup/{contextId}/category/{categoryId}/group/{groupId}/update": {
      "put": {
        "security": [
          {
            "OauthSecurity": ["read"]
          }
        ],
        "description": "update group equipements Group",
        "summary": "update group equipements Group",
        "tags": ["Equipements Group"],
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64."
            }
          },
          {
            "in": "path",
            "name": "categoryId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "path",
            "name": "groupId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["newNameGroup", "newNameColor"],
                "properties": {
                  "newNameGroup": {
                    "type": "string"
                  },
                  "newNameColor": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update Successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/api/v1/endPointsGroup/{contextId}/category/{categoryId}/group/{groupId}/update": {
      "put": {
        "security": [
          {
            "OauthSecurity": ["read"]
          }
        ],
        "description": "update group endPoints Group",
        "summary": "update group endPoints Group",
        "tags": ["EndPoints Group"],
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "path",
            "name": "categoryId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "path",
            "name": "groupId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["newNameGroup", "newNameColor"],
                "properties": {
                  "newNameGroup": {
                    "type": "string"
                  },
                  "newNameColor": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update Successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BasicNode": {
      "type": "object",
      "properties": {
        "dynamicId": {
          "type": "integer"
        },
        "staticId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
...
  },
...
}